requirements.txt
django~=3.3.1
djangorestframework~=3.11.0
djangorestframework-jwt~=1.11.0